Use packfiles for textures and use Sprite images
Added a class that can be run to take a directory of texture files and convert the images within it into a texture pack. A texture pack has a set of large images called pages, which contain the smaller images from the texture directory, and a .pack file, which describes how to find those smaller images in the larger image.

To add/remove an image and create a new texture pack:
  Place the file in the textures/ directory and run the com.splatform.util.textures.Packer class as a java application.

To use that image in the application:
  Sprite sprite = WorldRenderer.TEXTURES.createSprite("<image filename WITHOUT file extension>");